not-so-surgical extraction of blobs
