+++ yarn-project/circuit-types/src/interfaces/server_circuit_prover.ts
@@ -1,26 +1,28 @@
 import {
   type AVM_PROOF_LENGTH_IN_FIELDS,
   type AvmCircuitInputs,
-  type BaseOrMergeRollupPublicInputs,
   type BaseParityInputs,
-  type BlockMergeRollupInputs,
-  type BlockRootOrBlockMergePublicInputs,
-  type BlockRootRollupInputs,
-  type EmptyBlockRootRollupInputs,
   type KernelCircuitPublicInputs,
-  type MergeRollupInputs,
   type NESTED_RECURSIVE_PROOF_LENGTH,
   type ParityPublicInputs,
-  type PrivateBaseRollupInputs,
   type PrivateKernelEmptyInputData,
-  type PublicBaseRollupInputs,
   type RECURSIVE_PROOF_LENGTH,
   type RootParityInputs,
-  type RootRollupInputs,
-  type RootRollupPublicInputs,
   type TUBE_PROOF_LENGTH,
-  type TubeInputs,
 } from '@aztec/circuits.js';
+import {
+  BaseOrMergeRollupPublicInputs,
+  BlockMergeRollupInputs,
+  BlockRootOrBlockMergePublicInputs,
+  BlockRootRollupInputs,
+  EmptyBlockRootRollupInputs,
+  MergeRollupInputs,
+  PrivateBaseRollupInputs,
+  PublicBaseRollupInputs,
+  RootRollupInputs,
+  RootRollupPublicInputs,
+  TubeInputs,
+} from '@aztec/circuits.js/rollup';
 
 import type { Tx } from '../tx/tx.js';
 import { type ProofAndVerificationKey, type PublicInputsAndRecursiveProof } from './proving-job.js';
