+++ yarn-project/circuits.js/src/structs/blobs/index.ts
@@ -1,0 +1,2 @@
+export * from './blob_public_inputs.js';
+export * from './sponge_blob.js';
