+++ yarn-project/circuits.js/src/structs/blobs/sponge_blob.test.ts
@@ -1,8 +1,8 @@
 import { poseidon2Hash } from '@aztec/foundation/crypto';
 import { Fr } from '@aztec/foundation/fields';
 
-import { SPONGE_BLOB_LENGTH } from '../constants.gen.js';
-import { makeSpongeBlob } from '../tests/factories.js';
+import { SPONGE_BLOB_LENGTH } from '../../constants.gen.js';
+import { makeSpongeBlob } from '../../tests/factories.js';
 import { SpongeBlob } from './sponge_blob.js';
 
 describe('SpongeBlob', () => {
