+++ yarn-project/circuits.js/src/structs/blobs/sponge_blob.ts
