+++ yarn-project/circuits.js/src/structs/index.ts
@@ -2,7 +2,6 @@ export * from '@aztec/foundation/eth-address';
 export * from './avm/avm.js';
 export * from './avm/avm_accumulated_data.js';
 export * from './avm/avm_circuit_public_inputs.js';
-export * from './blob_public_inputs.js';
 export * from './call_context.js';
 export * from './client_ivc_proof.js';
 export * from './complete_address.js';
@@ -69,28 +68,8 @@ export * from './read_request.js';
 export * from './read_request_hints/index.js';
 export * from './recursive_proof.js';
 export * from './revert_code.js';
-export * from './rollup/append_only_tree_snapshot.js';
-export * from './rollup/avm_proof_data.js';
-export * from './rollup/base_or_merge_rollup_public_inputs.js';
-export * from './rollup/base_rollup_hints.js';
-export * from './rollup/block_merge_rollup.js';
-export * from './rollup/block_root_or_block_merge_public_inputs.js';
-export * from './rollup/constant_rollup_data.js';
-export * from './rollup/empty_block_root_rollup_inputs.js';
-export * from './rollup/block_root_rollup.js';
-export * from './rollup/merge_rollup.js';
-export * from './rollup/previous_rollup_block_data.js';
-export * from './rollup/previous_rollup_data.js';
-export * from './rollup/private_base_rollup_inputs.js';
-export * from './rollup/private_tube_data.js';
-export * from './rollup/public_base_rollup_inputs.js';
-export * from './rollup/public_tube_data.js';
-export * from './rollup/root_rollup.js';
-export * from './rollup/state_diff_hints.js';
-export * from './rollup/tube_inputs.js';
 export * from './rollup_validation_requests.js';
 export * from './scoped_key_validation_request_and_generator.js';
-export * from './sponge_blob.js';
 export * from './shared.js';
 export * from './state_reference.js';
 export * from './tree_leaf_read_request.js';
