+++ yarn-project/circuits.js/src/structs/partial_state_reference.ts
@@ -4,7 +4,7 @@ import { BufferReader, FieldReader, serializeToBuffer } from '@aztec/foundation/
 import { z } from 'zod';
 
 import { PARTIAL_STATE_REFERENCE_LENGTH } from '../constants.gen.js';
-import { AppendOnlyTreeSnapshot } from './rollup/append_only_tree_snapshot.js';
+import { AppendOnlyTreeSnapshot } from './trees/append_only_tree_snapshot.js';
 
 /**
  * Stores snapshots of trees which are commonly needed by base or merge rollup circuits.
