+++ yarn-project/circuits.js/src/structs/rollup/block_root_or_block_merge_public_inputs.ts
@@ -6,9 +6,9 @@ import { bufferToHex, hexToBuffer } from '@aztec/foundation/string';
 import { type FieldsOf } from '@aztec/foundation/types';
 
 import { AZTEC_MAX_EPOCH_DURATION } from '../../constants.gen.js';
-import { BlockBlobPublicInputs } from '../blob_public_inputs.js';
+import { BlockBlobPublicInputs } from '../blobs/blob_public_inputs.js';
 import { GlobalVariables } from '../global_variables.js';
-import { AppendOnlyTreeSnapshot } from './append_only_tree_snapshot.js';
+import { AppendOnlyTreeSnapshot } from '../trees/append_only_tree_snapshot.js';
 
 /**
  * Output of the block root and block merge rollup circuits.
