+++ yarn-project/circuits.js/src/structs/rollup/index.ts
@@ -1,0 +1,18 @@
+export * from './avm_proof_data.js';
+export * from './base_or_merge_rollup_public_inputs.js';
+export * from './base_rollup_hints.js';
+export * from './block_merge_rollup.js';
+export * from './block_root_or_block_merge_public_inputs.js';
+export * from './constant_rollup_data.js';
+export * from './empty_block_root_rollup_inputs.js';
+export * from './block_root_rollup.js';
+export * from './merge_rollup.js';
+export * from './previous_rollup_block_data.js';
+export * from './previous_rollup_data.js';
+export * from './private_base_rollup_inputs.js';
+export * from './private_tube_data.js';
+export * from './public_base_rollup_inputs.js';
+export * from './public_tube_data.js';
+export * from './root_rollup.js';
+export * from './state_diff_hints.js';
+export * from './tube_inputs.js';
