+++ yarn-project/circuits.js/src/structs/tree_snapshots.ts
@@ -4,7 +4,7 @@ import { BufferReader, FieldReader, serializeToBuffer } from '@aztec/foundation/
 import { inspect } from 'util';
 
 import { TREE_SNAPSHOTS_LENGTH } from '../constants.gen.js';
-import { AppendOnlyTreeSnapshot } from './rollup/append_only_tree_snapshot.js';
+import { AppendOnlyTreeSnapshot } from './trees/append_only_tree_snapshot.js';
 
 /**
  * Stores snapshots of all the trees but archive.
