+++ yarn-project/circuits.js/src/structs/trees/append_only_tree_snapshot.ts
