+++ yarn-project/circuits.js/src/structs/trees/index.ts
@@ -1,2 +1,3 @@
 export * from './nullifier_leaf.js';
 export * from './public_data_leaf.js';
+export * from './append_only_tree_snapshot.js';
