+++ yarn-project/prover-client/src/orchestrator/epoch-proving-state.ts
@@ -2,7 +2,6 @@ import { type MerkleTreeId } from '@aztec/circuit-types';
 import {
   type ARCHIVE_HEIGHT,
   type AppendOnlyTreeSnapshot,
-  type BlockRootOrBlockMergePublicInputs,
   Fr,
   type GlobalVariables,
   type L1_TO_L2_MSG_SUBTREE_SIBLING_PATH_LENGTH,
@@ -10,9 +9,9 @@ import {
   NUMBER_OF_L1_L2_MESSAGES_PER_ROLLUP,
   type Proof,
   type RecursiveProof,
-  type RootRollupPublicInputs,
   type VerificationKeyAsFields,
 } from '@aztec/circuits.js';
+import { BlockRootOrBlockMergePublicInputs, RootRollupPublicInputs } from '@aztec/circuits.js/rollup';
 import { padArrayEnd } from '@aztec/foundation/collection';
 import { type Tuple } from '@aztec/foundation/serialize';
