+++ yarn-project/prover-client/src/prover-agent/memory-proving-queue.ts
@@ -11,25 +11,27 @@ import {
 import type {
   AVM_PROOF_LENGTH_IN_FIELDS,
   AvmCircuitInputs,
-  BaseOrMergeRollupPublicInputs,
   BaseParityInputs,
+  KernelCircuitPublicInputs,
+  NESTED_RECURSIVE_PROOF_LENGTH,
+  ParityPublicInputs,
+  PrivateKernelEmptyInputData,
+  RECURSIVE_PROOF_LENGTH,
+  RootParityInputs,
+} from '@aztec/circuits.js';
+import {
+  BaseOrMergeRollupPublicInputs,
   BlockMergeRollupInputs,
   BlockRootOrBlockMergePublicInputs,
   BlockRootRollupInputs,
   EmptyBlockRootRollupInputs,
-  KernelCircuitPublicInputs,
   MergeRollupInputs,
-  NESTED_RECURSIVE_PROOF_LENGTH,
-  ParityPublicInputs,
   PrivateBaseRollupInputs,
-  PrivateKernelEmptyInputData,
   PublicBaseRollupInputs,
-  RECURSIVE_PROOF_LENGTH,
-  RootParityInputs,
   RootRollupInputs,
   RootRollupPublicInputs,
   TubeInputs,
-} from '@aztec/circuits.js';
+} from '@aztec/circuits.js/rollup';
 import { randomBytes } from '@aztec/foundation/crypto';
 import { AbortError, TimeoutError } from '@aztec/foundation/error';
 import { createLogger } from '@aztec/foundation/log';
