+++ yarn-project/pxe/src/pxe_service/error_enriching.ts
@@ -1,5 +1,8 @@
 import { type SimulationError, isNoirCallStackUnresolved } from '@aztec/circuit-types';
-import { AztecAddress, Fr, FunctionSelector, PUBLIC_DISPATCH_SELECTOR } from '@aztec/circuits.js';
+import { PUBLIC_DISPATCH_SELECTOR } from '@aztec/circuits.js/constants';
+import { FunctionSelector } from '@aztec/foundation/abi';
+import { AztecAddress } from '@aztec/foundation/aztec-address';
+import { Fr } from '@aztec/foundation/fields';
 import { type Logger } from '@aztec/foundation/log';
 import { resolveAssertionMessageFromRevertData, resolveOpcodeLocations } from '@aztec/simulator/errors';
